clase-8/project-clase/requirements.txt
flask
Flask-SQLAlchemy
psycopg2
pymysql
cryptography
flask_jwt_extended
flask_bcrypt
cerberus
python-dotenv